pyproject.toml
[project]
name = "verditai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "openpyxl>=3.1.5",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.22",
    "ruff>=0.15.6",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.41.0",
]
